Changes required Baking App

+++ app/src/main/java/com/example/computec/bakingapp/model/Ingredient.java
@@ -49,7 +49,7 @@ public class Ingredient implements Parcelable {
     }
 
     protected Ingredient(Parcel in) {
-        this.quantity = in.readInt();
+        this.quantity = in.readFloat();
         this.measure = in.readString();
         this.ingredient = in.readString();
     }
